# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Auth api",
        "version": "0.0.1"
    },
    "paths": {
        "/admin/api/index": {
            "get": {
                "operationId": "0decaddd0e529ecaf88281e92215043d",
                "responses": {
                    "200": {
                        "description": "Display a listing of projects."
                    }
                }
            }
        },
        "/api/dnf/subList": {
            "get": {
                "tags": [
                    "DNFM-GM"
                ],
                "summary": "获取子类列表",
                "operationId": "e3f507afc123236f067d68c403cb6960",
                "parameters": [
                    {
                        "name": "pid",
                        "in": "query",
                        "description": "父级id 1",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/dnf/propList": {
            "get": {
                "tags": [
                    "DNFM-GM"
                ],
                "summary": "获取物品列表",
                "operationId": "d4734dcf0edd7cc964fcb1b0bb5d2210",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页码 1",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "数量 10000",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "名称 点券",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "grade",
                        "in": "query",
                        "description": "等级 65",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type_id",
                        "in": "query",
                        "description": "种类id 1 材料 39 装备 23 消耗品 31 时装礼包",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "second_type_id",
                        "in": "query",
                        "description": "子类id 25",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/dnf/propNum": {
            "post": {
                "tags": [
                    "DNFM-GM"
                ],
                "summary": "发送物品",
                "operationId": "905089702b8dd02ba6b18ce935241125",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "用户名",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "code": {
                                        "description": "邀请码",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "id": {
                                        "description": "物品ID",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "num": {
                                        "description": "物品数量",
                                        "type": "integer",
                                        "format": "int32"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/dnf/multiSend": {
            "post": {
                "tags": [
                    "DNFM-GM"
                ],
                "summary": "批量发送物品",
                "operationId": "8543496216d0271114c7881e74d750f5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "username": {
                                        "description": "用户名",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "code": {
                                        "description": "邀请码",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "id": {
                                        "description": "物品ID",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "num": {
                                        "description": "物品数量",
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "page": {
                                        "description": "页码 1",
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "limit": {
                                        "description": "数量 10000",
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "title": {
                                        "description": "物品名称",
                                        "type": "string",
                                        "format": "text"
                                    },
                                    "grade": {
                                        "description": "等级",
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "type_id": {
                                        "description": "种类id 1 材料 39 装备 23 消耗品 31 时装礼包",
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "second_type_id": {
                                        "description": "子类id 25",
                                        "type": "integer",
                                        "format": "int32"
                                    },
                                    "filter": {
                                        "description": "物品名称过滤",
                                        "type": "string",
                                        "format": "text"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/excel/export": {
            "get": {
                "tags": [
                    "导出受保护且带水印的excel"
                ],
                "summary": "导出受保护且带水印的excel",
                "operationId": "864d6a90f2a45fa287aadf82617013b4",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/excel/watermark": {
            "get": {
                "tags": [
                    "导出受保护且带水印的excel"
                ],
                "summary": "水印图片生成",
                "operationId": "93251bb12b6d6ee36251f72771b09d18",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/excel/exportMulti": {
            "get": {
                "tags": [
                    "导出受保护且带水印的excel"
                ],
                "summary": "生成受保护且带水印的excel(数据量较大)",
                "operationId": "e1d5e485ac8d9dd2b5dc307040be3c6e",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "apiKey",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lines",
                        "in": "query",
                        "description": "测试函数 1000",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "测试文件名",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/excel/downloadMulti": {
            "get": {
                "tags": [
                    "导出受保护且带水印的excel"
                ],
                "summary": "导出受保护且带水印的excel",
                "operationId": "3b730c8abc92eb58b931f2906c97dff9",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "apiKey",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "lines",
                        "in": "query",
                        "description": "测试函数 1000",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "测试文件名",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/api/v1/admin/login/login": {
            "post": {
                "tags": [
                    "用户登录"
                ],
                "summary": "用户登录",
                "operationId": "0bfd53bef0dd9d44f594e72ce5636a1c",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "header",
                        "description": "apiKey",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "username",
                        "in": "query",
                        "description": "用户名",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "密码",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        }
    }
}